Texteditor superclass core
