+++ src/de/x28hd/tool/TextEditorPanel.java
@@ -1,7 +1,6 @@
 package de.x28hd.tool;
 
 
-import java.awt.BorderLayout;
 import java.awt.Desktop;
 import java.awt.Font;
 import java.awt.Toolkit;
@@ -24,67 +23,36 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javax.swing.Action;
-import javax.swing.JButton;
 import javax.swing.JComponent;
 import javax.swing.JEditorPane;
 import javax.swing.JMenuItem;
-import javax.swing.JPanel;
 import javax.swing.JPopupMenu;
-import javax.swing.JScrollPane;
 import javax.swing.TransferHandler;
 import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
-import javax.swing.event.CaretEvent;
-import javax.swing.event.CaretListener;
 import javax.swing.event.HyperlinkEvent;
 import javax.swing.event.HyperlinkListener;
 import javax.swing.event.UndoableEditEvent;
 import javax.swing.event.UndoableEditListener;
 import javax.swing.text.BadLocationException;
-import javax.swing.text.DefaultEditorKit;
-import javax.swing.text.EditorKit;
 import javax.swing.text.JTextComponent;
 import javax.swing.text.MutableAttributeSet;
-import javax.swing.text.StyledDocument;
 import javax.swing.text.StyledEditorKit;
 import javax.swing.text.html.HTML;
-import javax.swing.text.html.HTMLDocument;
 import javax.swing.text.html.HTMLEditorKit;
 import javax.swing.undo.UndoManager;
 import javax.swing.undo.UndoableEdit;
 
-public class TextEditorPanel extends JPanel implements ActionListener, UndoableEditListener, HyperlinkListener {
+public class TextEditorPanel extends TextEditorCore implements ActionListener, UndoableEditListener, HyperlinkListener {
 
-	public PresentationService controler;
-	public PresentationCore controCore;
-	boolean dumbCaller;	// to disable things temporarily
-	
-	private JEditorPane editorPane; 
-	EditorKit eki;;
-	HTMLDocument htmlDoc;
-	StyledDocument doc = null;
 	UndoManager undoManager = new UndoManager();
-    
-	private JScrollPane scrollPane;
-	JPanel toolbar;
-	StyledEditorKit.BoldAction boldAction = new StyledEditorKit.BoldAction();
-	String textToAdd = "";
-	CaretListener myCaretAdapter;
 	MyMouseMotionAdapter myMouseMotionAdapter = new MyMouseMotionAdapter();
-;
-	int selMark;
-	int selDot;
-	int myMark;
-	int myDot;
 	boolean isDirty = false;
 	boolean editableOrClickable = true; //  hyperlinks disabled
 	boolean hashesEnabled = false;
 	boolean tablet = false;
 	private static final long serialVersionUID = 1L;
-
 	int offset = -1;
-	String EndOfLineStringProperty = "NOCH NIX";
-	
 	boolean bundleInProgress = false;
 	boolean dragFake = false;
 	String myTransferable = "";
@@ -164,26 +132,8 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 		}
 	}
 
-	// special
-
-	class BoldSpecialActionAdapter implements ActionListener {
-		public void actionPerformed(ActionEvent a) {
-			if (a.getActionCommand().equals("AddToLabel")) {
-				boldAction.actionPerformed(a);
-				controCore.addToLabel(textToAdd);
-			}
-		}
-	}
-
 	public TextEditorPanel(Object caller) {
-		dumbCaller = !(caller instanceof PresentationService);
-		controCore = (PresentationCore) caller;
-		if (!dumbCaller) controler = (PresentationService) caller;
-		showDiag();
-		
-		setLayout(new BorderLayout());
-		editorPane = new JEditorPane();
-		
+		super(caller);
 		try {
 			UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
 		} catch (UnsupportedLookAndFeelException e) {
@@ -195,61 +145,7 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 		} catch (IllegalAccessException e) {
 			System.out.println("Error TE107 " + e);
 		}  
-
-		editorPane.getEditorKit().createDefaultDocument();
-		editorPane.setContentType("text/html");
-		editorPane.setText("<body><p style=\"margin-top: 0\">t</p></body>");
-		htmlDoc = (HTMLDocument) editorPane.getDocument();
 		editorPane.addHyperlinkListener(this);
-		editorPane.addCaretListener(myCaretAdapter);
-		doc = (StyledDocument) editorPane.getDocument();
-
-		editorPane.setEditable(true);		// ### false is required for hyperlinks to work
-		
-		scrollPane = new JScrollPane(editorPane);
-		add(scrollPane);
-
-//
-//		Toolbar for Bold, Italic, Underline and Bold Special
-		
-		toolbar = new JPanel();
-
-		// Bold: Action is a field because it is tapped in BoldSpecialActionAdapter
-		String boldActionCommand = (String) boldAction.getValue(Action.ACTION_COMMAND_KEY);	
-		JButton boldButton = new JButton("<html><body><b>B</b></body></html>");
-		boldButton.setToolTipText("Bold");
-		boldButton.setActionCommand(boldActionCommand);
-		boldButton.addActionListener(boldAction);
-		toolbar.add(boldButton);
-		
-		// Italic & Underline as expected
-    	Action italicAction = new StyledEditorKit.ItalicAction();
-		String italicActionCommand = (String) italicAction.getValue(Action.ACTION_COMMAND_KEY);	
-		JButton italicButton = new JButton("<html><body><i>I</i></body></html>");
-		italicButton.setToolTipText("Italic");
-		italicButton.setActionCommand(italicActionCommand);
-		italicButton.addActionListener(italicAction);
-		toolbar.add(italicButton);
-
-		Action underlineAction = new StyledEditorKit.UnderlineAction();
-		String underlineActionCommand = (String) underlineAction.getValue(Action.ACTION_COMMAND_KEY);
-		JButton underlineButton = new JButton("<html><body><u>U</u></body></html>");
-		underlineButton.setToolTipText("Underline");
-		underlineButton.setActionCommand(underlineActionCommand);
-		underlineButton.addActionListener(underlineAction);
-		toolbar.add(underlineButton);
-
-		// Bold special: "AddToLabel"
-		BoldSpecialActionAdapter boldSpecialActionListener = new BoldSpecialActionAdapter();
-		String boldSpecialActionCommand = "AddToLabel";
-
-		JButton boldSpecialButton = new JButton("<html><body><b>B+</b></body></html>");
-		boldSpecialButton.setToolTipText("Bold Special: bold and add the marked text to the item's label above and on the map");
-		boldSpecialButton.setActionCommand(boldSpecialActionCommand);
-		boldSpecialButton.addActionListener(boldSpecialActionListener);
-		toolbar.add(boldSpecialButton);
-		
-		add(toolbar, BorderLayout.SOUTH);
 
 //
 //		Accessory for Mouse: for Context Menu 
@@ -271,21 +167,6 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 				if (dragFake) selDot = 0;
 			}
 		});
-
-		myCaretAdapter = new MyCaretAdapter();
-		editorPane.addCaretListener(myCaretAdapter);
-	}
-	
-	class MyCaretAdapter implements CaretListener {
-
-		public void caretUpdate(CaretEvent e) {
-	          myDot = e.getDot();
-	          myMark = e.getMark();
-			selMark = e.getMark();
-			selDot = e.getDot();
-//			System.out.println("Dot = " + selDot + ", Mark = " + selMark );
-			processClicks();
-		}
 	}
 
 	class MyMouseMotionAdapter implements MouseMotionListener {
@@ -300,17 +181,7 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 	};
 	
 	public void processClicks() {
-		int selectedLen = selDot - selMark;
-		if (selectedLen < 0) {
-			selectedLen = - selectedLen;
-			selMark = selDot;
-		}
-		try {
-			textToAdd = editorPane.getText(selMark, selectedLen);
-		} catch (BadLocationException e1) {
-			System.out.println("Error TE103 " + e1);
-			textToAdd = "";
-		}
+		super.processClicks();
 		if (!textToAdd.isEmpty() && dragFake) {
 			//	enable drag
 			editorPane.removeCaretListener(myCaretAdapter);
@@ -484,23 +355,6 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
     }
     
 //
-//	Main Methods
-	
-	public void setText(String text) {
-		EndOfLineStringProperty = doc.getProperty(DefaultEditorKit.EndOfLineStringProperty).toString();
-		editorPane.setCaretPosition(selDot);
-		text = unwrap(text);
-		if (!text.startsWith("    <p") && !text.startsWith("<p")) {
-			text = "<p style=\"margin-top: 0\">" + text + "</p>"; 
-		}
-		editorPane.setText(text);
-	}
-	
-	public String getText() {
-		return editorPane.getText();
-	}
-	
-//
 //	Miscellaneous Accessories
 	
 	public void setSize(int size) {
@@ -508,21 +362,6 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 		editorPane.putClientProperty(JEditorPane.HONOR_DISPLAY_PROPERTIES, true);
 		editorPane.setFont(new Font("Serif", Font.PLAIN, size + 2));
 		repaint();
-	}
-	
-	public String unwrap(String text) {
-		// TODO create a regex
-		if (text.startsWith("\t<html>")) text = text.substring(7);	// NewStuff simple files
-		if (text.startsWith("<html>" + EndOfLineStringProperty)) text = text.substring(7);
-		if (text.startsWith("<html>")) text = text.substring(6);	// TreeImport
-		if (text.endsWith("</html>" + EndOfLineStringProperty)) text = text.substring(0, text.length() - 8);
-		if (text.startsWith("  <head>" + EndOfLineStringProperty)) text = text.substring(9);
-		if (text.startsWith("    " + EndOfLineStringProperty)) text = text.substring(5);
-		if (text.startsWith("  </head>" + EndOfLineStringProperty)) text = text.substring(10);
-		if (text.startsWith(EndOfLineStringProperty + "  </head>" + EndOfLineStringProperty)) text = text.substring(11);
-		if (text.startsWith("  <body>" + EndOfLineStringProperty)) text = text.substring(9);
-		if (text.endsWith("  </body>" + EndOfLineStringProperty)) text = text.substring(0, text.length() - 10);
-		return text;
 	}
 
 	public void toggleHyp() {
@@ -533,7 +372,6 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 	public void toggleTablet(boolean onOff) {
 		tablet = onOff;
 	}
-	
 	
 	public JTextComponent getTextComponent() {
 		return (JTextComponent) editorPane;
@@ -569,6 +407,7 @@ public class TextEditorPanel extends JPanel implements ActionListener, UndoableE
 		undoManager.addEdit(undoableEdit);
 		setDirty(true);
 	}
+	
 	public void showDiag() {
 		System.out.println("TE from dumb? " + dumbCaller);
 	}
